requirements.txt
asgiref==3.7.2
autopep8==2.0.4
certifi==2023.7.22
charset-normalizer==3.2.0
cloudinary==1.35.0
Django==4.2.5
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-jazzmin==2.6.0
django-threadlocals==0.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
idna==3.4
packaging==23.1
Pillow==10.0.1
psycopg2==2.9.8
pycodestyle==2.11.0
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.16
